Some progress has been made

+++ featureMedicine/src/main/java/cz/vvoleman/phr/featureMedicine/data/datasource/room/schedule/mapper/MedicineScheduleDataSourceModelToDataMapper.kt
@@ -26,7 +26,7 @@ class MedicineScheduleDataSourceModelToDataMapper(
         return MedicineScheduleDataSourceModel(
             id = model.id!!.toInt(),
             patientId = model.patient.id.toInt(),
-            medicineId = model.medicine.id.toInt(),
+            medicineId = model.medicine.id,
             createdAt = model.createdAt
         )
     }
